# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Youtube Video API",
        "description": "This is a Youtube Video API that allow users to retrieve generated data about youtube video(s) as a JSON format.",
        "contact": {
            "email": "n00201327@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/channel": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get details about your channel",
                "description": "Get everything related to your account such as name, subscriber, videos, comments etc.",
                "operationId": "5f0516e37fc40ea57a3d8e0f87462544",
                "responses": {
                    "200": {
                        "description": "Successful operation returns JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Channel"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Comment"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update all details in your channel",
                "description": "Update Channel details",
                "operationId": "44410b6f2543a98ffd159c723e86b5bd",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Update your channel name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Update your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Update your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Delete your channel",
                "description": "Delete your own channel from the database, including the videos, and comments by the channel",
                "operationId": "9c327252e6821d149ce77010c2dced64",
                "responses": {
                    "200": {
                        "description": "Channel has been successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update a specific detail in your channel",
                "description": "Update Channel details",
                "operationId": "8ca3e06b311fdf839c2a949473dc3b87",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Update your channel name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Update your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Update your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/channels": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Display channels retrieved from the database.",
                "description": "Displays all the youtube videos",
                "operationId": "be223b25163500bac9e4ec9d02150af6",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "orderBy allows data to be re-arrange in the order of descending order of id, name, created_at, and subscribers.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "random",
                        "in": "query",
                        "description": "retrieve data and pluck it in random order.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "subscribers",
                        "in": "query",
                        "description": "retrieve data by filtering the channels that has their subscribers below the specified threshold.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "retrieve data by filtering the channels that are made in the specified year.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hasVideos",
                        "in": "query",
                        "description": "retrieve data by filtering the channels that have videos.",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "yes",
                                "no"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Books in JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/channels/{id}/videos": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Display all the videos made by the channel by retreiving its ID.",
                "description": "Gets videos by its channel ID",
                "operationId": "c12cf4dc0da93cf936e12ea1b43bccdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/channels/{id}": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "Display the specified channel resource using its ID.",
                "description": "Gets a channel by its ID",
                "operationId": "ddb5c9deb38fca62c45ad8ac2ee85a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Channel id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/youtubeVideos/{id}": {
            "get": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Gets a youtube video by its ID",
                "description": "Retreive a youtube video specified by the ID parameter and displays it in a JSON format.",
                "operationId": "5b85c2afa115dc9c8692ed462f88004a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "video id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Update a youtube video by id",
                "description": "update and store the updated data to a specific youtube video in the database",
                "operationId": "put",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "duration",
                                    "likes",
                                    "dislikes",
                                    "views"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great book"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "60"
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "200"
                                    },
                                    "dislikes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "300"
                                    },
                                    "views": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/youtube_video"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid video ID"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Youtube Videos"
                ],
                "summary": "Delete a youtube video by its ID",
                "description": "Delete a youtube video specified by the ID parameter.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Video",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/successful_delete"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Channel": {
                "properties": {
                    "id": {
                        "description": "channel id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "name of the channel",
                        "type": "string",
                        "example": "James"
                    },
                    "subscribers": {
                        "description": "number of subscribers",
                        "type": "integer",
                        "example": "1100"
                    },
                    "created_at": {
                        "description": "the date when the channel has been created",
                        "type": "string",
                        "format": "date",
                        "example": "2019-10-06"
                    },
                    "hasVideo": {
                        "description": "the amount of video",
                        "type": "integer",
                        "format": "integer",
                        "example": "100"
                    },
                    "videos": {
                        "$ref": "#/components/schemas/youtube_video"
                    }
                },
                "type": "object"
            },
            "successful_delete": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The data has been successfully deleted"
                    },
                    "status": {
                        "type": "integer",
                        "example": "202"
                    }
                },
                "type": "object"
            },
            "auth": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "201"
                    },
                    "message": {
                        "type": "string",
                        "example": "Channel has been successfully registered"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|C7eINx2yy8TUM3PkWZXwt8OtN62ndluhpcYKjxIE"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "properties": {
                    "id": {
                        "description": "comment id",
                        "type": "integer",
                        "example": "1"
                    },
                    "text": {
                        "description": "comment texts",
                        "type": "text",
                        "example": "lorem ipsum"
                    },
                    "likes": {
                        "description": "number of likes",
                        "type": "integer",
                        "example": "1100"
                    },
                    "dislikes": {
                        "description": "number of dislikes",
                        "type": "integer",
                        "example": "90"
                    },
                    "commented_at": {
                        "description": "The date of the comment",
                        "type": "string",
                        "format": "string",
                        "example": "6 years ago"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}